[boostrap-5] Create types for Bootstrap 5
