+++ types/bootstrap/v4/package.json
@@ -1,0 +1,6 @@
+{
+    "private": true,
+    "dependencies": {
+        "popper.js": "^1.14.1"
+    }
+}
